+++ src/app/pages/all-tests/all-tests.component.html
@@ -1,5 +1,5 @@
 
 <pt-header></pt-header>
 <pt-create-new-block></pt-create-new-block>
-<pt-wraper></pt-wraper>
+<pt-tests-list></pt-tests-list>
