+++ src/app/pages/all-tests/create-new-block/create-new-block.component.html
@@ -1,9 +1,7 @@
 <div class="block">
     <div class="my-tests">
         <span class="my-tests-description">My tests</span>
-        <form action="/create-new-test">
-            <input type="submit" value="+ Create new" class="create-new"/>
-        </form>
+        <a class="create-new" routerLink="/create-new-test">+ Create new</a>
     </div>
     <div class="horizontal-line"></div>
 </div>
