+++ src/app/pages/all-tests/header/navigation-buttons/navigation-buttons.component.html
@@ -1,11 +1,5 @@
-<div >
-    <form action="/all-tests">
-        <input type="submit" value="All tests" class="black_text"/>
-    </form>
-    <form action="/all-activations">
-        <input type="submit" value="Activations" class="grey_text"/>
-    </form>
-    <form action="/all-respondents">
-        <input type="submit" value="All respondents" class="grey_text"/>
-    </form>
+<div class="menu">
+    <a class="menu-item" routerLinkActive="active" routerLink="/all-tests">All tests</a>
+    <a class="menu-item" routerLinkActive="active" routerLink="/all-activations">Activations</a>
+    <a class="menu-item" routerLinkActive="active" routerLink="/all-respondents">All responders</a>
 </div>
