+++ src/app/pages/all-tests/header/navigation-buttons/navigation-buttons.component.scss
@@ -1,28 +1,45 @@
 @import "../../../../shared/colors";
 
-.grey_text {
+.menu-item {
+  text-decoration: none;
   font-family: "Roboto", sans-serif;
   font-size: 20px;
   text-align: left;
   color: $silver;
+
+  &:not(:last-child) {
+    margin-right: 20px;
+  }
+
+  &.active {
+    color: $black;
+  }
 }
 
-.black_text {
+.menu {
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: center;
+  align-items: center;
+}
+
+.black-text {
   font-family: "Roboto", sans-serif;
   font-size: 20px;
   text-align: left;
   color: black;
 }
 
-form{
+form {
   display: inline-block;
   line-height: 61px;
   margin-left: 15px;
   margin-right: 15px;
 }
-input{
-  background:none;
-  border:none;
+
+input {
+  background: none;
+  border: none;
 }
 
 div {
