+++ src/app/pages/all-tests/tests-list/test-item/test-item.component.html
@@ -1,0 +1,13 @@
+<div class="test-item">
+    <span class="test-image"><img [src]="testImagePath"></span>
+    <span class="edited">Edited in {{test.lastEdited}} days ago</span>
+    <span class="test-title">{{test.name}}<br></span>
+    <span class="amount-of-questions">{{test.numberOfQuestions}} вопросов</span>
+    <div class="rectangles">&#9633;<br>&#9633;<br>&#9633;</div>
+    <div class="test-options">
+        <span>View Questions</span>
+        <span>View Activations</span>
+        <span>Edit</span>
+        <span>Activate</span>
+    </div>
+</div>
