+++ src/app/pages/all-tests/tests-list/test-item/test-item.component.scss
