+++ src/app/pages/all-tests/tests-list/tests-list.component.html
@@ -1,0 +1,3 @@
+<div>
+    <pt-test-item *ngFor="let test of tests" [test]="test"></pt-test-item>
+</div>
