+++ src/app/pages/all-tests/tests-list/tests-list.component.scss
