--- src/app/pages/all-tests/wraper/test-item/test-item.component.html
@@ -1,23 +1,0 @@
-<div class="test-item" *ngFor="let test of tests">
-    <span class="test-image"><img [src]="testImagePath"></span>
-    <span class="edited">Edited in {{test.lastEdited}} days ago</span>
-    <span class="test-title">{{test.name}}<br></span>
-    <span class="amount-of-questions">{{test.numberOfQuestions}} вопросов</span>
-    <div class="rectangles">&#9633;<br>&#9633;<br>&#9633;</div>
-    <div class="test-options">
-        <span>View Questions</span>
-        <span>View Activations</span>
-        <span>Edit</span>
-        <span>Activate</span>
-    </div>
-    <script>
-        function hideMainElements()
-        {
-            console.log('logg');
-            document.getElementsByClassName('test-title').setAttribute("display", "inline-block");
-            document.getElementsByClassName('test-title').setAttribute("display", "inline-block");
-            document.getElementsByClassName('amount-of-questions').setAttribute("display", "inline-block");
-            document.getElementsByClassName('rectangles').setAttribute("display", "inline-block");
-        }
-    </script>
-</div>
