--- src/app/pages/all-tests/wraper/wraper.component.html
@@ -1,3 +1,0 @@
-<div>
-    <pt-test-item></pt-test-item>
-</div>
