--- src/app/pages/all-tests/wraper/wraper.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'pt-wraper',
-  templateUrl: './wraper.component.html',
-  styleUrls: ['./wraper.component.scss']
-})
-export class WraperComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
